insulator_monitor/requirements.txt
flask
torch
transformers
Pillow